+++ crates/bevy_editor_pls/examples/empty.rs
@@ -4,6 +4,6 @@ use bevy_editor_pls::prelude::*;
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_plugin(EditorPlugin)
+        .add_plugin(EditorPlugin::new())
         .run();
 }