+++ crates/bevy_editor_pls/examples/ui.rs
@@ -10,7 +10,7 @@ use bevy_editor_pls::EditorPlugin;
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_plugin(EditorPlugin)
+        .add_plugin(EditorPlugin::new())
         // Only run the app when there is user input. This will significantly reduce CPU/GPU use.
         .insert_resource(WinitSettings::desktop_app())
         .add_startup_system(setup)